chore: active filter pills improvements (#10471)
* chore: add useActiveFilterPills hook

* chore: add TopArtworksContent prop

* chore: small improvements

* chore: rename prop
