+++ src/v2/Components/SavedSearchAlert/Components/ActiveFilterPills.tsx
@@ -1,53 +1,31 @@
 import React from "react"
-import { isArray } from "lodash"
-import {
-  ArtworkFilters,
-  initialArtworkFilterState,
-  useArtworkFilterContext,
-} from "v2/Components/ArtworkFilter/ArtworkFilterContext"
-import { extractPillsFromCriteria } from "../Utils/extractPills"
 import { SavedSearchAlertPills } from "./SavedSearchAlertPills"
 import { FilterPill } from "../types"
-import { getAllowedSearchCriteria } from "../Utils/savedSearchCriteria"
-import { DEFAULT_METRIC } from "v2/Components/ArtworkFilter/Utils/metrics"
-import { usePrepareFiltersForPills } from "v2/Components/ArtworkFilter/Utils/usePrepareFiltersForPills"
+import { useActiveFilterPills } from "../useActiveFilterPills"
+import { Flex } from "@artsy/palette"
 
-interface ActiveFilterPillsProps {
+export interface ActiveFilterPillsProps {
   defaultPills?: FilterPill[]
 }
 
+const PILL_HORIZONTAL_MARGIN_SIZE = 0.5
+
 export const ActiveFilterPills: React.FC<ActiveFilterPillsProps> = props => {
   const { defaultPills = [] } = props
-  const { aggregations, setFilter } = useArtworkFilterContext()
-  const filters = usePrepareFiltersForPills()
-  const criteria = getAllowedSearchCriteria(filters ?? {})
-  const metric = filters?.metric ?? DEFAULT_METRIC
-  const filterPills = extractPillsFromCriteria({
-    criteria,
-    aggregations,
-    metric,
-  })
-  const pills = [...defaultPills, ...filterPills]
+  const { pills, removePill } = useActiveFilterPills(defaultPills)
 
-  const removePill = (pill: FilterPill) => {
-    if (pill.isDefault) {
-      return
-    }
-
-    let filterValue = filters![pill.field]
-
-    if (isArray(filterValue)) {
-      filterValue = filterValue.filter(value => value !== pill.value)
-    } else {
-      filterValue = initialArtworkFilterState[pill.field]
-    }
-
-    setFilter(pill.field as keyof ArtworkFilters, filterValue)
-
-    if (pill.field === "artistIDs") {
-      setFilter("includeArtworksByFollowedArtists", false)
-    }
+  if (pills.length === 0) {
+    return null
   }
 
-  return <SavedSearchAlertPills items={pills} onDeletePress={removePill} />
+  return (
+    <Flex
+      flexWrap="wrap"
+      mx={-PILL_HORIZONTAL_MARGIN_SIZE}
+      mb={4}
+      data-testid="artworkGridFilterPills"
+    >
+      <SavedSearchAlertPills items={pills} onDeletePress={removePill} />
+    </Flex>
+  )
 }
